Really move mcl_stairs_stone_slab_top.png into mcl_core
This is a fix for #105 where I accidentally moved it to the wrong place.
Sorry.

mcl_core uses this texture for Polished Stone (mcl_core:stone_smooth),
which means mcl_core had an undeclared dependency on mcl_stairs.

Since mcl_stairs depends on mcl_core, it makes sense to move the texture
to mcl_core.